--- charts/core/src/Tooltip/TooltipDataList/index.ts
@@ -1,1 +1,0 @@
-export { TooltipDataList } from './TooltipDataList';
