+++ charts/core/src/Tooltip/TooltipSeriesList/TooltipSeriesList.types.ts
@@ -1,0 +1,11 @@
+import { ReactNode } from 'react';
+
+import { AxisFormatterCallbackParams } from '../Tooltip.types';
+
+export interface TooltipSeriesListProps {
+  params: AxisFormatterCallbackParams;
+  seriesValueFormatter?: (value: string | number | Date) => string | ReactNode;
+  seriesNameFormatter?: (seriesName: string) => string | ReactNode;
+  sortDirection?: 'asc' | 'desc';
+  sortValue?: 'name' | 'value';
+}
