+++ charts/core/src/Tooltip/TooltipSeriesList/index.ts
@@ -1,0 +1,2 @@
+export { TooltipSeriesList } from './TooltipSeriesList';
+export type { TooltipSeriesListProps } from './TooltipSeriesList.types';
