+++ charts/core/src/Tooltip/TooltipSeriesListItem/TooltipSeriesListItem.types.ts
@@ -1,0 +1,11 @@
+import { TooltipSeriesListProps } from '../TooltipSeriesList/TooltipSeriesList.types';
+
+export interface TooltipSeriesListItemProps
+  extends Pick<
+    TooltipSeriesListProps,
+    'seriesValueFormatter' | 'seriesNameFormatter'
+  > {
+  seriesName?: string;
+  data: [string | number, string | number | Date];
+  color: string;
+}
