+++ charts/core/src/Tooltip/TooltipSeriesListItem/index.ts
@@ -1,0 +1,2 @@
+export { TooltipSeriesListItem } from './TooltipSeriesListItem';
+export type { TooltipSeriesListItemProps } from './TooltipSeriesListItem.types';
